Export styles for both Terms of Service and Privacy Policy from single file
